pyproject.toml
[tool.poetry]
name = "drf-tuts"
version = "0.1.0"
description = ""
authors = ["baabbii <yohanesalemu0069@gmail.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asgiref = "3.8.1"
django = "5.1.1"
django-extensions = "3.2.3"
djangorestframework = "3.15.2"
pillow = "10.4.0"
sqlparse = "0.5.1"
tzdata = "2024.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
